requirements.txt
pystray~=0.19.3
Pillow~=9.0.1